backend/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
pandas==2.2.1
openpyxl==3.1.2
djangorestframework==3.15.1
